implimenting authorization using gate

+++ app/Http/Controllers/QuestionsController.php
@@ -70,6 +70,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'acess denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -82,6 +85,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'acess denied');
+        } 
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has been updated");
@@ -95,6 +102,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, 'acess denied');
+        } 
+
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted successfully");
